+++ src/utils/load-commands.js
@@ -18,6 +18,8 @@ function requireCommands () {
     ls: require('../files-regular/ls'),
     lsReadableStream: require('../files-regular/ls-readable-stream'),
     lsPullStream: require('../files-regular/ls-pull-stream'),
+    refs: require('../files-regular/refs'),
+    refsPullStream: require('../files-regular/refs-pull-stream'),
 
     // Files MFS (Mutable Filesystem)
     files: require('../files-mfs'),
@@ -50,7 +52,6 @@ function requireCommands () {
     key: require('../key'),
     log: require('../log'),
     mount: require('../mount'),
-    refs: require('../refs'),
     repo: require('../repo'),
     stop: require('../stop'),
     shutdown: require('../stop'),
